+++ migrations/basics/rt_config.php
@@ -10,48 +10,49 @@
  * Config migration: extension configuration values
  */
 
-namespace avathar\recenttopics\migrations\basics;
+namespace avathar\recenttopicsav\migrations\basics;
 
 class rt_config extends \phpbb\db\migration\migration
 {
-	static public function depends_on()
+	public function effectively_installed()
 	{
-		return array(
-			'\avathar\recenttopics\migrations\basics\rt_perms',
-		);
+		return $this->config->offsetExists('rt_number');
+	}
+
+	public static function depends_on()
+	{
+		return ['\avathar\recenttopicsav\migrations\basics\rt_perms'];
 	}
 
 	public function update_data()
 	{
-		return array(
-			array('config.add', array('rt_number', '5')),
-			array('config.add', array('rt_page_number', 0)),
-			array('config.add', array('rt_page_numbermax', '10')),
-			array('config.add', array('rt_anti_topics', 0)),
-			array('config.add', array('rt_parents', 1)),
-			array('config.add', array('rt_index', 1)),
-			array('config.add', array('rt_min_topic_level', 0)),
-			array('config.add', array('rt_on_newspage', 0)),
-			array('config.add', array('rt_sort_start_time', 0)),
-			array('config.add', array('rt_unread_only', 0)),
-			array('config.add', array('rt_location', 'RT_TOP')),
-		);
+		return [
+			['config.add', ['rt_number', '5']],
+			['config.add', ['rt_page_number', 0]],
+			['config.add', ['rt_page_numbermax', '10']],
+			['config.add', ['rt_anti_topics', 0]],
+			['config.add', ['rt_parents', 1]],
+			['config.add', ['rt_index', 1]],
+			['config.add', ['rt_min_topic_level', 0]],
+			['config.add', ['rt_sort_start_time', 0]],
+			['config.add', ['rt_unread_only', 0]],
+			['config.add', ['rt_location', 'RT_TOP']],
+		];
 	}
 
 	public function revert_data()
 	{
-		return array(
-			array('config.remove', array('rt_number')),
-			array('config.remove', array('rt_page_number')),
-			array('config.remove', array('rt_page_numbermax')),
-			array('config.remove', array('rt_anti_topics')),
-			array('config.remove', array('rt_parents')),
-			array('config.remove', array('rt_index')),
-			array('config.remove', array('rt_min_topic_level')),
-			array('config.remove', array('rt_on_newspage')),
-			array('config.remove', array('rt_sort_start_time')),
-			array('config.remove', array('rt_unread_only')),
-			array('config.remove', array('rt_location')),
-		);
+		return [
+			['config.remove', ['rt_number']],
+			['config.remove', ['rt_page_number']],
+			['config.remove', ['rt_page_numbermax']],
+			['config.remove', ['rt_anti_topics']],
+			['config.remove', ['rt_parents']],
+			['config.remove', ['rt_index']],
+			['config.remove', ['rt_min_topic_level']],
+			['config.remove', ['rt_sort_start_time']],
+			['config.remove', ['rt_unread_only']],
+			['config.remove', ['rt_location']],
+		];
 	}
 }